refactor: ♻️ refactor types and plugin context
create (more) atomic type definitions and extract a plugin context to pass to classes.
